requirements.txt
flask==3.0.3
werkzeug==3.0.3
anthropic==0.28.0
python-dotenv==1.0.1
openai==1.35.0
gunicorn==22.0.0
